+++ src/main/java/com/knu/cse/board/domain/Category.java
@@ -1,4 +1,4 @@
-package com.knu.cse.Board.domain;
+package com.knu.cse.board.domain;
 
 public enum Category {
     FREE,QNA
